backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt[crypto]==5.4.0
django-cors-headers==4.4.0
psycopg[binary]==3.2.11
